% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbkreg.R
\name{add_where_clause}
\alias{add_where_clause}
\title{Add a WHERE clause to a FROM statement safely}
\usage{
add_where_clause(from_statement, where_clause)
}
\description{
Add a WHERE clause to a FROM statement safely
}
\keyword{internal}
