% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{apply_ridge_penalty}
\alias{apply_ridge_penalty}
\title{Apply ridge penalty to XtX}
\usage{
apply_ridge_penalty(XtX, ridge, penalize_idx = NULL)
}
\description{
Apply ridge penalty to XtX
}
\keyword{internal}
