% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbkreg.R
\name{build_grid_source}
\alias{build_grid_source}
\title{Build grid source (CTE or temp table)}
\usage{
build_grid_source(conn, backend, grid_df, xvars, cte_max = 500L)
}
\description{
Build grid source (CTE or temp table)
}
\keyword{internal}
