% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbdml.R
\name{coef.dbdml}
\alias{coef.dbdml}
\title{Extract coefficients from dbdml objects}
\usage{
\method{coef}{dbdml}(object, ...)
}
\arguments{
\item{object}{A `dbdml` object.}

\item{...}{Unused.}
}
\description{
Extract coefficients from dbdml objects
}
