% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{coef.dbreg}
\alias{coef.dbreg}
\title{Extract coefficients from dbreg objects}
\usage{
\method{coef}{dbreg}(object, fes = FALSE, ...)
}
\arguments{
\item{object}{a model of class `dbreg` produced by the [dbreg()] function}

\item{fes}{Should the fixed effects be included? Default is `FALSE`.}

\item{...}{Additional arguments (currently unused).}
}
\description{
Extract coefficients from dbreg objects
}
