% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbbin.R
\name{compute_bin_geometry}
\alias{compute_bin_geometry}
\title{Compute bin geometry from binned data}
\usage{
compute_bin_geometry(binned_data, x_name)
}
\description{
Compute bin geometry from binned data
}
\keyword{internal}
