% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbdml.R
\name{confint.dbdml}
\alias{confint.dbdml}
\title{Confidence intervals for dbdml objects}
\usage{
\method{confint}{dbdml}(object, parm, level = 0.95, ...)
}
\arguments{
\item{object}{A `dbdml` object.}

\item{parm}{Subset of parameters to return.}

\item{level}{Confidence level.}

\item{...}{Unused.}
}
\description{
Confidence intervals for dbdml objects
}
