% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbreg.R
\name{count_nested_fe_levels}
\alias{count_nested_fe_levels}
\title{Count levels of FE variables nested within cluster variable}
\usage{
count_nested_fe_levels(conn, from_statement, fe, cluster_var)
}
\description{
For ssc = "nested", we exclude FE levels from K if the FE is nested within
the cluster variable (i.e., each FE value belongs to exactly one cluster).
}
\keyword{internal}
