% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_bins.R
\name{create_binned_data}
\alias{create_binned_data}
\title{Build bin assignment CTE and fetch binned data into R}
\usage{
create_binned_data(inputs)
}
\description{
Build bin assignment CTE and fetch binned data into R
}
\keyword{internal}
