% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbreg-package.R
\docType{package}
\name{dbreg-package}
\alias{dbreg-package}
\title{dbreg: Fast Regressions on Database Backends}
\description{
Leverages database backends to run regressions on very large datasets, which may not fit into R's memory. Various acceleration strategies---e.g., Wong et al. (2021) \doi{10.48550/arXiv.2102.11297} and Arkhangelsky & Imbens (2024) \url{https://doi.org/10.1093/restud/rdad089}---allow for highly efficient computation, while robust standard errors are computed from sufficient statistics.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://grantmcdermott.com/dbreg/}
  \item Report bugs at \url{https://github.com/grantmcdermott/dbreg/issues}
}

}
\author{
\strong{Maintainer}: Grant McDermott \email{gmcd@amazon.com} (\href{https://orcid.org/0000-0001-7883-8573}{ORCID})

Authors:
\itemize{
  \item James Brand \email{jamesbrandecon@gmail.com} (\href{https://orcid.org/0000-0002-1663-5583}{ORCID})
}

Other contributors:
\itemize{
  \item Apoorva Lal \email{lal.apoorva@gmail.com} (\href{https://orcid.org/0000-0002-3697-614X}{ORCID}) [contributor]
  \item Laurent Berge \email{laurent.berge@u-bordeaux.fr} [contributor]
}

}
\keyword{internal}
