% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbreg.R
\name{dbreg_alternating_projections}
\alias{dbreg_alternating_projections}
\title{Alternating projections (AP) for exact two-way FE demeaning}
\usage{
dbreg_alternating_projections(
  conn,
  from_statement,
  fe,
  yvar,
  xvars_sql,
  xvar_names,
  weights,
  cluster_var = NULL,
  verbose = FALSE,
  max_iter = getOption("dbreg.ap_max_iter", 100L),
  tol = getOption("dbreg.ap_tol", 1e-10)
)
}
\description{
Alternating projections (AP) for exact two-way FE demeaning
}
\keyword{internal}
