% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbreg.R
\name{dbreg_create_temp_table_as}
\alias{dbreg_create_temp_table_as}
\title{Create a temp table from a SELECT query}
\usage{
dbreg_create_temp_table_as(conn, table_name, select_sql, backend)
}
\description{
Create a temp table from a SELECT query
}
\keyword{internal}
