% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbreg.R
\name{dbreg_is_balanced_panel}
\alias{dbreg_is_balanced_panel}
\title{Check if a two-way panel is balanced}
\usage{
dbreg_is_balanced_panel(conn, from_statement, fe)
}
\description{
Check if a two-way panel is balanced
}
\keyword{internal}
