% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbreg.R
\name{dbreg_temp_table_name}
\alias{dbreg_temp_table_name}
\title{Generate a temp table name with optional SQL Server prefix}
\usage{
dbreg_temp_table_name(base_name, backend)
}
\description{
Generate a temp table name with optional SQL Server prefix
}
\keyword{internal}
