% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{env2env}
\alias{env2env}
\title{Copy named values between environments}
\usage{
env2env(source, target, keys = NULL)
}
\arguments{
\item{source}{Source environment to read from}

\item{target}{Target environment to write to}

\item{keys}{Character vector of names to copy. If NULL, copies all.}
}
\description{
Copy named values between environments
}
\keyword{internal}
