% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbbin.R
\name{eval_poly}
\alias{eval_poly}
\title{Evaluate polynomial at a point}
\usage{
eval_poly(beta, u, degree)
}
\arguments{
\item{beta}{Coefficient vector (length degree+1)}

\item{u}{Centered x value}

\item{degree}{Polynomial degree}
}
\value{
Scalar polynomial value
}
\description{
Evaluate polynomial at a point
}
\keyword{internal}
