% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbbin.R
\name{eval_se}
\alias{eval_se}
\title{Evaluate standard error of polynomial at a point}
\usage{
eval_se(V, u, degree)
}
\arguments{
\item{V}{Covariance matrix of coefficients}

\item{u}{Centered x value}

\item{degree}{Polynomial degree}
}
\value{
Scalar standard error
}
\description{
Evaluate standard error of polynomial at a point
}
\keyword{internal}
