% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbbin.R
\name{execute_bins_design_ols}
\alias{execute_bins_design_ols}
\title{Execute bins estimation using design matrix / OLS approach}
\usage{
execute_bins_design_ols(inputs)
}
\description{
Execute bins estimation using design matrix / OLS approach
}
\keyword{internal}
