% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbbin.R
\name{execute_bins_moments_kkt}
\alias{execute_bins_moments_kkt}
\title{Execute bins estimation using moment-based KKT approach}
\usage{
execute_bins_moments_kkt(inputs)
}
\description{
Execute bins estimation using moment-based KKT approach
}
\keyword{internal}
