% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbreg.R
\name{execute_demean_strategy}
\alias{execute_demean_strategy}
\title{Execute demean strategy (1+ fixed effects)}
\usage{
execute_demean_strategy(inputs)
}
\description{
Double demeaning / within estimator. Gives identical coefficients to 
fixed effects regression.
}
\keyword{internal}
