% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbreg.R
\name{execute_mundlak_strategy}
\alias{execute_mundlak_strategy}
\title{Execute mundlak strategy (1-2 fixed effects)}
\usage{
execute_mundlak_strategy(inputs, weight_col = NULL)
}
\description{
Execute mundlak strategy (1-2 fixed effects)
}
\keyword{internal}
