% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbbinsreg.R
\name{execute_separate_binsreg}
\alias{execute_separate_binsreg}
\title{Execute separate binsreg models for points and line}
\usage{
execute_separate_binsreg(inputs)
}
\description{
When points and line have different (p,s) parameters, we need to fit two separate
models following binsreg's approach. This function:
1. Fits a model with points parameters (points.p, points.s) for data.dots
2. Fits a model with line parameters (line.p, line.s) for data.line
3. Combines results into single dbbinsreg output
}
\keyword{internal}
