% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_model_matrix.R
\name{expand_term}
\alias{expand_term}
\title{Expand a single term into SQL expressions}
\usage{
expand_term(term, col_info, expand)
}
\description{
Expand a single term into SQL expressions
}
\keyword{internal}
