% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_model_matrix.R
\name{expand_variable}
\alias{expand_variable}
\title{Expand a single variable into SQL expression(s)}
\usage{
expand_variable(var, col_info, expand, in_interaction)
}
\description{
Expand a single variable into SQL expression(s)
}
\keyword{internal}
