% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbbinsreg.R
\name{extract_knots}
\alias{extract_knots}
\title{Extract interior knots from bin geometry}
\usage{
extract_knots(geo)
}
\arguments{
\item{geo}{Data frame with bin geometry (must have x_right column)}
}
\value{
Numeric vector of K = B-1 interior knots
}
\description{
Extract interior knots from bin geometry
}
\keyword{internal}
