% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_model_matrix.R
\name{get_column_info}
\alias{get_column_info}
\title{Get column types and factor levels from database}
\usage{
get_column_info(conn, table, term_labels)
}
\description{
Get column types and factor levels from database
}
\keyword{internal}
