% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy.R
\name{glance.dbreg}
\alias{glance.dbreg}
\title{glance `dbreg` objects}
\usage{
\method{glance}{dbreg}(x, ...)
}
\arguments{
\item{x}{a model of class `dbreg` produced by the [dbreg()] function}

\item{...}{Additional arguments to glancing method.}
}
\description{
glance `dbreg` objects
}
