% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gof.R
\name{gof}
\alias{gof}
\title{Calculate goodness-of-fit metrics for dbreg objects}
\usage{
gof(object, ...)
}
\arguments{
\item{object}{A `dbreg` object.}

\item{...}{Additional arguments (currently unused)}
}
\value{
Named vector with r2, adj_r2, and rmse
}
\description{
Calculate goodness-of-fit metrics for dbreg objects
}
\keyword{internal}
