% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbkreg.R
\name{kernel_weight_sql_1d}
\alias{kernel_weight_sql_1d}
\title{Kernel weight SQL expression (1D)}
\usage{
kernel_weight_sql_1d(dx, h, kernel)
}
\description{
Kernel weight SQL expression (1D)
}
\keyword{internal}
