% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbkreg.R
\name{kernel_weight_sql_mdim}
\alias{kernel_weight_sql_mdim}
\title{Kernel weight SQL expression (multi-D product kernel)}
\usage{
kernel_weight_sql_mdim(xvars, bandwidth, kernel)
}
\description{
Kernel weight SQL expression (multi-D product kernel)
}
\keyword{internal}
