% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbbinsreg.R
\name{lagrange_interp_3pt}
\alias{lagrange_interp_3pt}
\title{Lagrange interpolation through 3 points}
\usage{
lagrange_interp_3pt(x_seq, x_pts, y_pts)
}
\arguments{
\item{x_seq}{Vector of x values to interpolate at}

\item{x_pts}{3-element vector of x control points}

\item{y_pts}{3-element vector of y control points}
}
\value{
Vector of interpolated y values
}
\description{
Lagrange interpolation through 3 points
}
\keyword{internal}
