% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.r
\name{plot.dbbinsreg}
\alias{plot.dbbinsreg}
\alias{tinyplot.dbbinsreg}
\title{Plot method for dbbinsreg objects}
\usage{
\method{plot}{dbbinsreg}(
  x,
  type = NULL,
  ci = TRUE,
  cb = TRUE,
  line = TRUE,
  lty = 1,
  theme = "basic",
  ...
)

\method{tinyplot}{dbbinsreg}(
  x,
  type = NULL,
  ci = TRUE,
  cb = TRUE,
  line = TRUE,
  lty = 1,
  theme = "basic",
  ...
)
}
\arguments{
\item{x}{A \code{dbbinsreg} object}

\item{type}{The type of plot. If \code{NULL} (the default), then the type will be
inferred based on the underlying object (e.g, \code{"pointrange"} for points with
confidence intervals).}

\item{ci}{Logical. Show confidence intervals for dots? Default is \code{TRUE.}}

\item{cb}{Logical. Show confidence bands as a ribbon? Default is \code{TRUE} if
available in the object.}

\item{line}{Logical. Show the line overlay if available? Default is \code{TRUE.}}

\item{lty}{Integer or character string. Line type for line overlay.}

\item{theme}{Character string. One of the valid plot themes supported by
\code{\link[tinyplot]{tinytheme}}. The default \code{"basic"} theme is a light
adaptation of the standard base graphics aesthetic, featuring filled points
and a background grid. Various other themes are supported (e.g., \code{"clean"},
\code{"minimal"}, \code{"classic"}, etc.), while passing \code{NULL} switches the theme off
entirely.}

\item{...}{Additional arguments passed to \code{\link[tinyplot]{tinyplot}},
e.g. \code{main}, \code{sub}, \code{file}, etc.}
}
\description{
Visualizes binned regression results from \code{\link{dbbinsreg}}.
Plots dots at bin means with optional confidence intervals and/or
confidence bands, and optionally overlays a smooth line if computed.
Uses tinyplot for rendering but works with both \code{plot()} and
\code{tinyplot()} generics.
}
\examples{
\dontrun{
#
## In-memory data ----

# Like `dbreg`, we can pass in-memory R data frames to an ephemeral DuckDB
# connection via the `data` argument. 

# Canonical binscatter: bin means (default)
dbbinsreg(weight ~ Time, data = ChickWeight, nbins = 10)

# You can pass additional plotting arguments via ... to (tiny)plot.dbbinsreg
dbbinsreg(weight ~ Time, data = ChickWeight, nbins = 10,
          main = "A simple binscatter example", theme = "clean")

# Alternatively, save the object and plot later
bs = dbbinsreg(weight ~ Time, data = ChickWeight, nbins = 10, plot = FALSE)
plot(bs, main = "Same example, different theme", theme = "classic")

# Piecewise linear (p = 1), no smoothness (s = 0)
dbbinsreg(weight ~ Time, data = ChickWeight, nbins = 10, points = c(1, 0))

# Piecewise linear (p = 1) with continuity (s = 1)
dbbinsreg(weight ~ Time, data = ChickWeight, nbins = 10, points = c(1, 1))

# With line overlay for smooth visualization
dbbinsreg(weight ~ Time, data = ChickWeight, nbins = 10, points = c(1, 1), line = TRUE)

# Different line smoothness to points
dbbinsreg(weight ~ Time, data = ChickWeight, nbins = 10, points = c(0, 0), line = c(1, 1))

# With uniform confidence bands (much greater uncertainty)
set.seed(99)
dbbinsreg(weight ~ Time, data = ChickWeight, nbins = 10, cb = TRUE)

# Accounting for Diet "fixed effects" helps to resolve the situation
# (We'll also add a line and change the theme for a nicer plot)
dbbinsreg(weight ~ Time | Diet, data = ChickWeight, nbins = 10, cb = TRUE,
          line = c(1, 1), theme = "clean")

#
## DBI connection ----

library(DBI)
con = dbConnect(duckdb::duckdb())
dbWriteTable(con, "cw", as.data.frame(ChickWeight))

dbbinsreg(weight ~ Time | Diet, conn = con, table = "cw", nbins = 10,
          theme = "clean")
# etc.

# See ?dbreg for more connection examples

# Clean up
dbDisconnect(con)
}

}
