% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.r
\name{plot.dbkreg}
\alias{plot.dbkreg}
\alias{tinyplot.dbkreg}
\title{Plot method for dbkreg objects}
\usage{
\method{plot}{dbkreg}(x, type = "l", theme = "basic", ...)

\method{tinyplot}{dbkreg}(x, type = "l", theme = "basic", ...)
}
\arguments{
\item{x}{A \code{dbkreg} object.}

\item{type}{Plot type passed to \code{\link[tinyplot]{tinyplot}}. Default is \code{"l"}.}

\item{theme}{Character string giving the tinyplot theme.}

\item{...}{Additional arguments passed to \code{\link[tinyplot]{tinyplot}}.}
}
\description{
Visualizes kernel regression results from \code{\link{dbkreg}}.
}
