% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats-methods.R
\name{predict.dbforest}
\alias{predict.dbforest}
\title{Predict method for dbforest objects}
\usage{
\method{predict}{dbforest}(object, newdata, individual = FALSE, ...)
}
\arguments{
\item{object}{A `dbforest` object.}

\item{newdata}{Data frame for predictions.}

\item{individual}{Logical. If `TRUE`, also return individual tree
predictions.}

\item{...}{Additional arguments (unused).}
}
\description{
Predict method for dbforest objects
}
