% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats-methods.R
\name{predict.dbtree}
\alias{predict.dbtree}
\title{Predict method for dbtree objects}
\usage{
\method{predict}{dbtree}(object, newdata, ...)
}
\arguments{
\item{object}{A `dbtree` object.}

\item{newdata}{Data frame for predictions.}

\item{...}{Additional arguments (unused).}
}
\description{
Predict method for dbtree objects
}
