% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbbin.R
\name{print.dbbin}
\alias{print.dbbin}
\title{Print method for dbbin objects}
\usage{
\method{print}{dbbin}(x, ...)
}
\arguments{
\item{x}{A dbbin object}

\item{...}{Additional arguments passed to print}
}
\description{
Print method for dbbin objects
}
