% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.dbbinsreg}
\alias{print.dbbinsreg}
\title{Print method for dbbinsreg objects (binsreg-compatible format)}
\usage{
\method{print}{dbbinsreg}(x, ...)
}
\arguments{
\item{x}{A dbbinsreg object}

\item{...}{Additional arguments passed to print}
}
\description{
Print method for dbbinsreg objects (binsreg-compatible format)
}
