% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.dbforest}
\alias{print.dbforest}
\title{Print method for dbforest objects}
\usage{
\method{print}{dbforest}(x, ...)
}
\arguments{
\item{x}{A `dbforest` object.}

\item{...}{Additional arguments (unused).}
}
\description{
Print method for dbforest objects
}
