% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbivreg.R
\name{print.dbivreg}
\alias{print.dbivreg}
\title{Print method for dbivreg objects}
\usage{
\method{print}{dbivreg}(x, ...)
}
\arguments{
\item{x}{A \code{dbivreg} object.}

\item{...}{Additional unused arguments.}
}
\description{
Print method for \code{dbivreg} objects.
}
