% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.dbkreg}
\alias{print.dbkreg}
\title{Print method for dbkreg objects}
\usage{
\method{print}{dbkreg}(x, ...)
}
\arguments{
\item{x}{A `dbkreg` object.}

\item{...}{Additional arguments passed to print.}
}
\description{
Print method for dbkreg objects
}
