% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.dbreg}
\alias{print.dbreg}
\title{Print method for dbreg objects}
\usage{
\method{print}{dbreg}(x, fe = FALSE, ...)
}
\arguments{
\item{x}{`dbreg` object.}

\item{fe}{Should the fixed effects be displayed? Default is `FALSE`.}

\item{...}{Other arguments passed to \code{\link[base]{print}}. Currently
unused, except to capture superseded arguments.}
}
\description{
Print method for dbreg objects
}
