% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.dbtree}
\alias{print.dbtree}
\title{Print method for dbtree objects}
\usage{
\method{print}{dbtree}(x, ...)
}
\arguments{
\item{x}{A dbtree object}

\item{...}{Additional arguments (unused)}
}
\description{
Print method for dbtree objects
}
