% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ridge_penalty_indices}
\alias{ridge_penalty_indices}
\title{Identify ridge-penalized columns in a design matrix}
\usage{
ridge_penalty_indices(X, fe_vars = NULL, penalize_intercept = FALSE)
}
\description{
Identify ridge-penalized columns in a design matrix
}
\keyword{internal}
