% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbkreg.R
\name{sample_from_statement}
\alias{sample_from_statement}
\title{Sample rows from a FROM clause (backend-specific)}
\usage{
sample_from_statement(conn, backend, from_statement, n)
}
\description{
Sample rows from a FROM clause (backend-specific)
}
\keyword{internal}
