% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sql_ntile}
\alias{sql_ntile}
\title{Get NTILE window function expression}
\usage{
sql_ntile(x_name, n_bins)
}
\arguments{
\item{x_name}{Column name to partition by}

\item{n_bins}{Number of bins}
}
\value{
SQL NTILE expression
}
\description{
Get NTILE window function expression
}
\keyword{internal}
