% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{standardize_coef_names}
\alias{standardize_coef_names}
\title{Convert internal interaction notation to standard R notation}
\usage{
standardize_coef_names(x)
}
\description{
Convert internal interaction notation to standard R notation
}
\keyword{internal}
