% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy.R
\name{tidy.dbreg}
\alias{tidy.dbreg}
\title{tidy `dbreg` objects}
\usage{
\method{tidy}{dbreg}(x, conf.int = FALSE, conf.level = 0.95, fes = FALSE, ...)
}
\arguments{
\item{x}{a model of class `dbreg` produced by the [dbreg()] function}

\item{conf.int}{Logical indicating whether to include confidence intervals.
Default is `FALSE`.}

\item{conf.level}{Confidence level for intervals. Default is 0.95.}

\item{fes}{Should the fixed effects be tidied too? Default is `FALSE`.}

\item{...}{Additional arguments to tidying method.}
}
\description{
tidy `dbreg` objects
}
