% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbdml.R
\name{vcov.dbdml}
\alias{vcov.dbdml}
\title{Variance-covariance matrix for dbdml objects}
\usage{
\method{vcov}{dbdml}(object, ...)
}
\arguments{
\item{object}{A `dbdml` object.}

\item{...}{Unused.}
}
\description{
Variance-covariance matrix for dbdml objects
}
