% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbbin.R
\name{add_basis_columns}
\alias{add_basis_columns}
\title{Add centered polynomial basis to binned data}
\usage{
add_basis_columns(binned_data, geo, x_name, degree)
}
\description{
Add centered polynomial basis to binned data
}
\keyword{internal}
