% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats-methods.R
\name{coef.dbreg}
\alias{coef.dbreg}
\title{Extract coefficients from dbreg objects}
\usage{
\method{coef}{dbreg}(object, fe = FALSE, ...)
}
\arguments{
\item{object}{A `dbreg` object.}

\item{fe}{Should the fixed effects be included? Default is `FALSE`.}

\item{...}{Additional arguments. Currently unused, except to capture
superseded arguments.}
}
\description{
Extract coefficients from dbreg objects
}
\examples{
mod = dbreg(Temp ~ Wind | Month, data = airquality)

# coefficients
coef(mod)
coef(mod, fe = TRUE)  # include fixed effects

# confidence intervals
confint(mod)

# variance-covariance matrix
vcov(mod)

# predictions
head(predict(mod, newdata = airquality))
head(predict(mod, newdata = airquality, interval = "confidence"))
}
