% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbreg.R
\name{compute_meat_sql}
\alias{compute_meat_sql}
\title{Compute HC1 meat matrix via SQL}
\usage{
compute_meat_sql(
  conn,
  cte_sql,
  vars,
  yvar,
  betahat,
  is_athena = FALSE,
  var_suffix = "_tilde",
  cte_name = "demeaned",
  has_intercept = FALSE,
  vars_sql = NULL
)
}
\description{
Compute HC1 meat matrix via SQL
}
\keyword{internal}
