% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confint.R
\name{confint.dbreg}
\alias{confint.dbreg}
\title{Confidence intervals for dbreg objects}
\usage{
\method{confint}{dbreg}(object, parm, level = 0.95, fes = FALSE, ...)
}
\arguments{
\item{object}{a model of class `dbreg` produced by the [dbreg()] function.}

\item{parm}{a specification of which parameters are to be given confidence
intervals, either a vector of numbers or a vector of names. If missing,
all parameters are considered.}

\item{level}{the confidence level required. Default is 0.95.}

\item{fes}{Should the fixed effects be included? Default is `FALSE`.}

\item{...}{Additional arguments (currently unused).}
}
\description{
Confidence intervals for dbreg objects
}
