% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_bins.R
\name{construct_output}
\alias{construct_output}
\title{Construct output tibble from fitted model}
\usage{
construct_output(inputs, fit, geo)
}
\description{
Construct output tibble from fitted model
}
\keyword{internal}
