% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbbin.R
\name{construct_output_from_moments}
\alias{construct_output_from_moments}
\title{Construct output from constrained solution}
\usage{
construct_output_from_moments(beta, geo, degree, smooth, partition)
}
\arguments{
\item{beta}{Coefficient vector (length B*(degree+1))}

\item{geo}{Tibble with bin geometry}

\item{degree}{Polynomial degree}

\item{smooth}{Smoothness level}

\item{partition}{Partition type}
}
\value{
Tibble with dbbin output structure
}
\description{
Construct output from constrained solution
}
\keyword{internal}
