% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbbin.R
\name{execute_separate_binsreg}
\alias{execute_separate_binsreg}
\title{Execute separate binsreg models for dots and line}
\usage{
execute_separate_binsreg(inputs)
}
\description{
When dots and line have different (p,s) parameters, we need to fit two separate
models following binsreg's approach. This function:
1. Fits a model with dots parameters (dots.p, dots.s) for data.dots
2. Fits a model with line parameters (line.p, line.s) for data.line
3. Combines results into single dbbin output
}
\keyword{internal}
