% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{predict.dbreg}
\alias{predict.dbreg}
\title{Predict method for dbreg objects}
\usage{
\method{predict}{dbreg}(
  object,
  newdata = NULL,
  interval = c("none", "confidence", "prediction"),
  level = 0.95,
  ...
)
}
\arguments{
\item{object}{a model of class `dbreg` produced by the [dbreg()] function}

\item{newdata}{Data frame for predictions. Required since dbreg does not
retain the original data.}

\item{...}{Additional arguments (currently unused).}
}
\description{
Predict method for dbreg objects
}
