% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.dbreg}
\alias{print.dbreg}
\title{Print method for dbreg objects}
\usage{
\method{print}{dbreg}(x, fe = FALSE, ...)
}
\arguments{
\item{x}{\code{dbreg} object.}

\item{fe}{Should the fixed effects be displayed? Default is \code{FALSE}.}

\item{...}{Other arguments passed to \code{\link[base]{print}}. Currently
unused, except to capture superseded arguments.}
}
\description{
Print method for dbreg objects
}
\examples{
mod = dbreg(Temp ~ Wind | Month, data = airquality)
# mod # same as below
print(mod)
print(mod, fe = TRUE)  # include fixed effects
}
