% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbreg.R
\name{process_dbreg_inputs}
\alias{process_dbreg_inputs}
\title{Process and validate dbreg inputs}
\usage{
process_dbreg_inputs(
  fml,
  conn,
  table,
  data,
  path,
  weights,
  vcov,
  cluster,
  ssc,
  strategy,
  sql_only,
  data_only,
  compress_ratio,
  compress_nmax,
  ridge,
  drop_missings,
  verbose
)
}
\description{
Process and validate dbreg inputs
}
\keyword{internal}
