% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidiers.R
\name{tidiers}
\alias{tidiers}
\alias{tidy.dbreg}
\alias{glance.dbreg}
\title{Tidiers for `dbreg` objects}
\usage{
\method{tidy}{dbreg}(x, conf.int = FALSE, conf.level = 0.95, fe = FALSE, ...)

\method{glance}{dbreg}(x, ...)
}
\arguments{
\item{x}{a model of class `dbreg` produced by the \code{\link[dbreg]{dbreg}}
function.}

\item{conf.int}{Logical indicating whether to include confidence intervals.
Default is `FALSE`.}

\item{conf.level}{Confidence level for intervals. Default is 0.95.}

\item{fe}{Should the fixed effects be tidied too? Default is `FALSE`.}

\item{...}{Additional arguments to tidying method. Currently unused except to
handle superseded arguments.}
}
\description{
Provides \code{broom::tidy} and \code{broom::glance} methods for
"dbreg" objects.
}
\examples{
mod = dbreg(Temp ~ Wind | Month, data = airquality)
tidy(mod, conf.int = TRUE)
tidy(mod, conf.int = TRUE, fe = TRUE)
glance(mod)
}
