% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy.R
\name{tidy.dbreg}
\alias{tidy.dbreg}
\title{tidy `dbreg` objects}
\usage{
\method{tidy}{dbreg}(x, fes = FALSE, ...)

\method{tidy}{dbreg}(x, fes = FALSE, ...)
}
\arguments{
\item{x}{a model of class `dbreg` produced by the [dbreg()] function}

\item{fes}{Should the fixed effects be tidied too? Default is `FALSE`.}

\item{...}{Additional arguments to tidying method.}
}
\description{
tidy `dbreg` objects

tidy `dbreg` objects
}
