% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats-methods.R
\name{vcov.dbreg}
\alias{vcov.dbreg}
\title{Variance-covariance matrix for dbreg objects}
\usage{
\method{vcov}{dbreg}(object, ...)
}
\arguments{
\item{object}{A `dbreg` object.}

\item{...}{Additional arguments (currently unused).}
}
\description{
Variance-covariance matrix for dbreg objects
}
\examples{
mod = dbreg(Temp ~ Wind | Month, data = airquality)

# coefficients
coef(mod)
coef(mod, fe = TRUE)  # include fixed effects

# confidence intervals
confint(mod)

# variance-covariance matrix
vcov(mod)

# predictions
head(predict(mod, newdata = airquality))
head(predict(mod, newdata = airquality, interval = "confidence"))
}
