% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confint.R
\name{vcov.dbreg}
\alias{vcov.dbreg}
\title{Variance-covariance matrix for dbreg objects}
\usage{
\method{vcov}{dbreg}(object, ...)
}
\arguments{
\item{object}{a model of class `dbreg` produced by the [dbreg()] function}

\item{...}{Additional arguments (currently unused).}
}
\description{
Variance-covariance matrix for dbreg objects
}
